refactor: LottoGame 컨트롤러 분리
